LatinIME: add a "Download Dictionary" step to the setup wizard
this commit does the following changes.

first we convert the debug class into a real activity, make it launchable
via a custom action.

then we add another step to the Setup Wizard, to download dictionaries and
launch the new activity from it.

this makes sure the user is not confused about the recent changes of LatinIME.
(to remind you, we removed the prebuilt dictionaries and offer them as optional
download inside LatinIME)

Change-Id: If996046b4b664b38dc9de673a74ae592cf6df702